Job list: apply job, the data will add to the Job database.
Registered list: the applied job, the data will add to the Registered job database

When at the job list pressed apply button, and then go to the registered list, the applied job will shown in real time.

When at the registered list pressed cancel button, and then the data will be removed and update the recyclerview in real time.